fix: 페이지 새로고침시 url 정리
